deploy script now checks if there are uncommited git changes before publishing
